monitor/oldFrontend/backend/requirements.txt
Flask~=2.0
paho-mqtt==1.6.1
requests==2.26.0
# urllib3==1.26.7
python-dateutil==2.8.2
ntplib==0.4.0
gunicorn==20.1.0


# pip3 install -r requirements.txt
